% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{birthrate}
\alias{birthrate}
\title{Vital statistics by country}
\format{A data frame with 97 rows and 4 variables:
\describe{
  \item{birth}{Birth rate (units unspecified)}
  \item{death}{Death rate}
  \item{infant}{Infant mortality rate}
  \item{country}{Name of country, no spaces}
}}
\source{
\url{https://support.sas.com/documentation/cdl/en/statug/63033/HTML/default/viewer.htm#statug_aceclus_sect003.htm}, from 1990 United Nations Demographic Yearbook.
}
\usage{
data(birthrate)
}
\description{
1990-era birth rates, death rates and infant mortality rates for
97 countries world-wide
}
\keyword{datasets}

