% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Ovary cancer and factors affecting survival}
\format{A data frame with 16 rows and 5 variables:
\describe{
  \item{stage}{Stage of cancer, early or advanced}
  \item{operation}{Type of operation, radical or limited}
  \item{xray}{Whether X-ray treatment was used, Yes or No}
  \item{survival}{Whether the patient survived, Yes or No}
  \item{freq}{Number of women observed with that combination of factors}
}}
\source{
Andersen, E. (2012), "Introduction to the Statistical Analysis of Categorical
Data", Springer.

Neave and Worthington (1988) "Distribution-Free Tests", Unwin-Hyman, page 110
}
\usage{
data(cancer)

data(cars)
}
\description{
A 1973 10-year retrospective study of 299 women with cancer of the ovary. Data
in modelling-friendly format.

8 Japanese and 12 British models of cars were assessed
and the percentage of cars of each model that break down
within the first year was recorded.
}
\keyword{datasets}

