% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{chainsaw}
\alias{chainsaw}
\title{Angle of deflection of chainsaw kickback}
\format{A data frame with 5 rows and 4 columns:
\describe{
  \item{A}{Deflection angles for chainsaw model A}
  \item{B}{Deflection angles for chainsaw model B}
  \item{C}{Deflection angles for chainsaw model C}
  \item{D}{Deflection angles for chainsaw model D}
}}
\source{
Dowdy, Wearden and Chilko (2004) "Statistics for Research", Wiley, p. 274
}
\usage{
data(chainsaw)
}
\description{
When a hand-held chainsaw begins to cut a piece of wood, there is a deflection
called "kickback" which can cause physiological stress to the user. Four different
manufacturers' models were tested (five different saws randomly chosen for each model).
A standard 3-inch board was used, and the angle of deflection for each chainsaw (in
degrees) was measured.
}
\keyword{datasets}

