% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{cov}
\alias{cov}
\title{A correlation matrix}
\format{A data frame with 3 rows and 3 columns, labelled V1, V2, V3.}
\usage{
data(cov)
}
\description{
An example correlation matrix of 3 variables to be used as input for principal
components or factor analysis.
}
\keyword{datasets}

