% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{cube}
\alias{cube}
\title{A cube}
\format{A data frame with 8 rows and 8 columns, with rows and columns labelled a-h to represent the
8 vertices of a cube and entries the (rounded) distances between those vertices of the cube}
\source{
Created by Ken Butler
}
\usage{
data(cube)
}
\description{
Distance matrix (as data frame) representing a 1x1x1 cube
}
\keyword{datasets}

