% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{curvy}
\alias{curvy}
\title{Curved relationship}
\format{10 rows of 2 variables:

\describe{
  \item{xx}{Explanatory variable for relationship}
  \item{yy}{Response variable for relationship}
}}
\source{
Made up by Ken Butler.
}
\usage{
data(curvy)
}
\description{
Relationship between two variables that is a curve rather than a straight line
}
\keyword{datasets}

