% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{dancing}
\alias{dancing}
\title{Still dancing?}
\format{12 rows of 4 variables:

\describe{
  \item{Months}{Number of months a woman was observed after starting dance lessons}
  \item{Quit}{Whether or not the woman quit dancing: 1=yes, 0=no}
  \item{Treatment}{Whether or not the women went to an event with live dancing: 1=yes, 0=no}
  \item{Age}{Woman's age at start of dance lessons}
}}
\source{
Tabachnick and Fidell (2013), "Using Multivariate Statistics", Pearson, 6th ed., page 515
}
\usage{
data(dancing)
}
\description{
Assessment of whether women who took dancing lessons are still dancing
some number of months later
}
\keyword{datasets}

