% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{dogs}
\alias{dogs}
\title{Effect of trimethaphan and morphine on histamine levels in dogs}
\format{8 rows of 6 variables:

\describe{
  \item{Drug}{Morphine or Trimethaphan}
  \item{x}{Whether the dog had depleted histamine levels before the experiment, always N (No) here.}
  \item{lh0}{Log-histamine level at time 0}
  \item{lh1}{Log-histamine level at time 1}
  \item{lh3}{Log-histamine level at time 3}
  \item{lh5}{Log-histamine level at time 5}
}}
\source{
Cole, J. W. L., & Grizzle, J. E. (1966). Applications of multivariate analysis of variance
to repeated measurements experiments. Biometrics, 810-828, obtained from
\url{http://www.sfu.ca/sasdoc/sashtml/stat/chap30/sect56.htm}.
}
\usage{
data(dogs)
}
\description{
Eight dogs are randomly assigned to receive either morphine or trimethophan. Their log-histamine
levels were measured at 0, 1, 3 and 5 minutes after injection of the drug. (In the original study,
sixteen dogs were used, eight of which had depleted histamine levels before the experiment. Those
dogs are omitted here.)
}
\keyword{datasets}

