% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docco.R
\docType{data}
\name{drp}
\alias{drp}
\title{Effect of reading activities on Degree of Reading Power scores}
\format{44 rows of 2 variables:

\describe{
  \item{group}{Treatment group (t), including activities, on control group (c) without activities}
  \item{score}{Degree of Reading Power score at end of study}
}}
\source{
Moore, David S., and George P. McCabe (1989). Introduction to the Practice of Statistics.
 Description above taken from \url{https://www.statcrunch.com/5.0/shareddata.php?keywords=CLASSROOM},
 data set "Improving Reading Ability".
}
\usage{
data(drp)
}
\description{
Results of an experiment to test whether directed reading activities in the classroom help
 elementary school students improve aspects of their reading ability. A treatment class of 21
 third-grade students participated in these activities for eight weeks, and a control class of 23
 third-graders followed the same curriculum without the activities. After the eight-week
 period, students in both classes took a Degree of Reading Power (DRP) test which measures the
 aspects of reading ability that the treatment is designed to improve.
}
\keyword{datasets}

