% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docce.R
\docType{data}
\name{ecg}
\alias{ecg}
\title{Effect of overweight and smoking on coronary heart disease}
\format{8 rows of 4 variables:

\describe{
  \item{ecg}{whether the electrocardiagram was normal or abnormal}
  \item{bmi}{Indication of normal weight or overweight by body mass index}
  \item{smoke}{Smoker, yes or no}
  \item{count}{Frequency of combination of other variables}
}}
\source{
\url{http://www.oxfordjournals.org/our_journals/tropej/online/ma_chap14.pdf}, from which
above explanation also taken.
}
\usage{
data(ecg)
}
\description{
In a hospital accident and emergency service 176 subjects who
attended for acute chest pain were enrolled in a study. Of these 71 had abnormal
electrocardiograms and in the case of 105 it was normal. Of those with
abnormal electrocardiograms, 57 were overweight as judged by their body
mass index, and 14 were normal. By comparison out of the 105 subjects
with normal electrocardiograms 40 were overweight and 65 normal.
In the first group of 71 subjects with abnormal electrocardiograms, out of
the 57 overweight subjects 47 were smokers and 10 non-smokers.
Amongst the 14 with normal weights 8 were smokers and 6 non-smokers.
In the second group of 105 with normal electrocardiograms out of the 40
overweight subjects 25 were smokers and 15 non-smokers. Amongst the
65 with normal weights 35 were smokers and 30 non-smokers.
The investigators wish to assess the contribution that overweight and
smoking make to coronary artery disease.
}
\keyword{datasets}

