% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docce.R
\docType{data}
\name{europe}
\alias{europe}
\title{Road distances between European cities}
\format{16 rows of 17 variables:

\describe{
  \item{X}{City names (playing the role of row names)}
  \item{Amsterdam...Vienna}{Columns of distances from Amsterdam...Vienna}
}}
\source{
\url{http://www.engineeringtoolbox.com/driving-distances-d_1029.html}, selection of the
cities there.
}
\usage{
data(europe)
}
\description{
Matrix (as data frame) of road distances (in km) between European cities
}
\keyword{datasets}

