% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docce.R
\docType{data}
\name{eyewear}
\alias{eyewear}
\title{Eyewear and gender}
\format{2 rows of 4 variables:

\describe{
  \item{gender}{Male or female}
  \item{contacts}{Count of people of that gender wearing contact lenses}
  \item{glasses}{Count of people of that gender wearing glasses}
  \item{none}{Count of people of that gender wearing no eyewear}
}}
\source{
Pfenning, "Elementary Statistics: Looking at the Big Picture",  Nelson, p. 153
}
\usage{
data(eyewear)
}
\description{
Contingency table of people by gender wearing glasses, contact lenses or no eyewear
}
\keyword{datasets}

