% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docce.R
\docType{data}
\name{fertilizer}
\alias{fertilizer}
\title{Effect of fertilizer on seed yield and weight}
\format{8 rows of 3 variables:

\describe{
  \item{fert}{high or low nitrogen level fertilizer}
  \item{yield}{Total seed per plant (grams)}
  \item{weight}{Total weight of 100 randomly selected seeds per plant (milligrams)}
}}
\source{
\url{http://www.imm.dtu.dk/~perbb/MAS/ST116/module04/index.html}
}
\usage{
data(fertilizer)
}
\description{
Completely-randomized agricultural field trial assessing effects of fertilizer (low or high) on
seed yield and seed weight in (unnamed) plants.
}
\keyword{datasets}

